requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.9.1
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
biopython==1.84
bitsandbytes==0.45.0
black==24.10.0
blake3==1.0.0
bleach==6.2.0
blinker==1.9.0
blis==1.1.0
boto3==1.35.83
botocore==1.35.83
brotli==1.1.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.8.1
confection==0.1.5
crontab==1.0.1
cryptography==43.0.3
cut-cross-entropy==24.12.2
cymem==2.0.10
dataclasses-json==0.5.7
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.0.0
docstring-parser==0.16
einops==0.8.0
email-validator==2.2.0
eval-type-backport==0.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
flask==3.1.0
flask-rq2==18.3
fqdn==1.5.1
freezegun==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
healthcheck==1.3.3
hf-transfer==0.1.8
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipyplot==1.1.2
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json-repair==0.32.0
json5==0.10.0
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
langcodes==3.5.0
language-data==1.3.0
lark==1.2.2
libdnf5==0.0.1a1
litellm==1.52.12
lm-format-enforcer==0.10.9
loguru==0.7.2
lorax @ file:///home/grahama/dev/lorax/dist/lorax-42.4-py3-none-any.whl#sha256=cd7a6eb784c50c296b6f00d47cf78ca34c948b318b64aaad541a2735beb0ab00
lorax-client==0.6.3
marisa-trie==1.2.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.2
marshmallow-enum==1.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.5.1
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
mypy==1.13.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.1
notebook-shim==0.2.4
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.0
opencv-python-headless==4.10.0.84
orjson==3.10.12
outlines==0.1.11
outlines-core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
peft==0.14.0
pexpect==4.9.0
pfzy==0.3.4
phonenumbers==8.13.52
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
predibase==2024.12.3
predibase-api==2024.12.3
preshed==3.0.9
prettytable==3.12.0
progress-table==0.1.26
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
prompt-toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybind11==2.13.6
pycares==4.5.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pygments==2.18.0
pyisemail==2.0.1
pyjwt==2.10.1
pynacl==1.5.0
pyperclip==1.9.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-arango==7.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
python-rapidjson==1.20
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
raven==6.10.0
ray==2.40.0
redis==5.2.1
redis-sentinel-url==1.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rq==2.0.0
rq-dashboard==0.8.2.2
rq-scheduler==0.14.0
runpod==1.7.4
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
selinux==0.3.0
semantic-version==2.10.0
send2trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.6.0
setuptools-scm==8.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.18.1
thinc==8.3.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
tqdm-loggable==0.2
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
tritonclient==2.41.1
trl==0.13.0
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
typing-inspect==0.9.0
tyro==0.9.4
tzdata==2024.2
ujson==5.10.0
unsloth==2024.12.8
unsloth-zoo==2024.12.3
uri-template==1.3.0
urllib3==1.26.12
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.6.5
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xformers==0.0.28.post3
xgrammar==0.1.7
xxhash==3.5.0
yarl==1.18.3
yaspin==3.1.0
zipp==3.21.0
